requirements.txt
agent-framework
python-dotenv>=1.0.0
azure-identity>=1.15.0
httpx>=0.27.0
